Stateless Animation API
"Relnote: Introduced low level stateless animation APIs. These APIs
describe the primitive animations that have been the building blocks of
the higher level APIs such as animate() and @Composable Transition

Specifically, the new APIs include:
1) New AnimationSpec interface that describes the configuration/formula of an
animation of vectors. The implementations of this API include RepeatableSpec,
KeyframeSpec, SpringSpec, SnapSpec, and TweenSpec.

2) An interface named Animation that caches start/end conditions along with
an AnimationSpec instance for more convenient APIs.

3) FloatAnimationSpec, and FloatDecayAnimationSpec interfaces that describes
animation configurations specifically for floats. These float specs can be
vectorized into AnimationSpec of vector type.
"

BUG: 158189074
BUG: 157588223
Test: unit tests and benchmark added in this CL

+++ ui/ui-animation/src/androidTest/java/androidx/ui/animation/CrossfadeTest.kt
@@ -51,7 +51,7 @@ class CrossfadeTest {
                 Text(if (it) First else Second)
             }
         }
-        composeTestRule.clockTestRule.advanceClock(DefaultDuration.toLong())
+        composeTestRule.clockTestRule.advanceClock(DefaultDuration)
 
         findByText(First).assertExists()
     }
@@ -70,7 +70,7 @@ class CrossfadeTest {
                 }
             }
         }
-        composeTestRule.clockTestRule.advanceClock(DefaultDuration.toLong())
+        composeTestRule.clockTestRule.advanceClock(DefaultDuration)
 
         runOnIdleCompose {
             showFirst = false
@@ -78,7 +78,7 @@ class CrossfadeTest {
 
         waitForIdle()
 
-        composeTestRule.clockTestRule.advanceClock(DefaultDuration.toLong())
+        composeTestRule.clockTestRule.advanceClock(DefaultDuration)
 
         runOnIdleCompose {
             assertTrue(disposed)
